+++ mistelaflask/views/admin_views/admin_view_guests.py
@@ -35,7 +35,7 @@ class AdminViewGuests(AdminViewBase):
 
         _events = models.Event.query.all()
         guest_list = []
-        for _guest in models.User.query.filter_by(admin=False):
+        for _guest in models.User.query.filter_by():
             _invitations = []
             for _event in _events:
                 gi = GuestInvitation()
@@ -159,11 +159,11 @@ class AdminViewGuests(AdminViewBase):
             if _invited:
                 _invitations.append(_event)
 
-        for name in _names_list.splitlines():
+        for _idx, name in enumerate(_names_list.splitlines()):
             _full_name = name.strip()
             _username = _full_name[0:3] + _full_name[-4:-1]
             if models.User.query.filter_by(username=_username).first():
-                _username = _username + str(len(models.User.query.all()))
+                _username = _username + str(_idx)
             _user = models.User(
                 username=_username,
                 name=_full_name,