+++ mistelaflask/views/guest_views/guest_view_invitations.py
@@ -1,7 +1,16 @@
 from __future__ import annotations
 
-from flask import Blueprint, redirect, render_template, request, url_for
+from flask import (
+    Blueprint,
+    current_app,
+    flash,
+    redirect,
+    render_template,
+    request,
+    url_for,
+)
 from flask_login import current_user, login_required
+from flask_mail import Mail, Message
 
 from mistelaflask import db, models
 from mistelaflask.views.guest_views.guest_view_protocol import GuestViewProtocol
@@ -9,6 +18,7 @@ from mistelaflask.views.guest_views.guest_view_protocol import GuestViewProtocol
 
 class GuestViewInvitations(GuestViewProtocol):
     view_name: str = "invitations"
+    _mail: Mail = None
 
     def _render_guest_view_template(self, template_name: str, **context):
         return render_template(f"guest/" + template_name, **context)
@@ -42,6 +52,12 @@ class GuestViewInvitations(GuestViewProtocol):
             (ui.response, ui.n_adults, ui.n_children, ui.n_babies, ui.remarks)
             for ui in user_invitations
         ]
+        if not _all_values:
+            flash(
+                f"An error occured while loading your invitation. Please contact us if the problem persists.",
+                category="danger",
+            )
+            return redirect(url_for("main.index"))
         _list_values = list(zip(*_all_values))
         _assisting = all(_list_values[0])
         if not _assisting:
@@ -81,6 +97,7 @@ class GuestViewInvitations(GuestViewProtocol):
         _remarks = request.form.get(f"remarks", "").strip()
         if _remarks:
             _remarks = _remarks.splitlines()
+        _email = request.form.get("email", "")
         for _idx, _inv in enumerate(
             models.UserEventInvitation.query.filter_by(guest_id=current_user.id)
         ):
@@ -94,6 +111,27 @@ class GuestViewInvitations(GuestViewProtocol):
                 _inv.remarks = ""
             db.session.add(_inv)
             db.session.commit()
+        with current_app.app_context():
+            _admin_mail = current_app.config.get("MAIL_USERNAME")
+            msg = Message(
+                subject=f"RSVP from {current_user.name}",
+                sender=_admin_mail,
+                recipients=[
+                    _admin_mail,
+                ],  # replace with your email for testing
+                body="""
+                Response: {}
+                Adults: {}
+                Children: {}
+                Babies: {}
+                Remarks: {}
+                Email: {}
+                """.format(
+                    _response, _n_adults, _n_children, _n_babies, _remarks, _email
+                ),
+            )
+            _mail = Mail(current_app)
+            _mail.send(msg)
         return redirect(url_for("main.invitations_list"))
 
     @login_required